requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
alembic==1.9.2
aniso8601==8.0.0
anyio==3.6.2
astunparse==1.6.3
bcrypt==4.0.1
beautifulsoup4==4.11.1
cachetools==5.2.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
ci-info==0.3.0
click==8.1.3
colorama==0.4.4
configobj==5.0.6
configparser==5.3.0
contourpy==1.0.7
cryptography==38.0.1
cycler==0.11.0
distlib==0.3.4
dlib==19.24.0
etelemetry==0.3.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fer==22.4.0
filelock==3.7.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.1
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==3.0.2
flatbuffers==23.1.4
fonttools==4.38.0
frontend==0.0.3
future==0.18.3
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.51.1
h11==0.14.0
h5py==3.7.0
hid==1.0.4
httplib2==0.21.0
idna==3.3
importlib-metadata==6.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
looseversion==1.0.3
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
MouseInfo==0.1.3
mtcnn==0.1.1
mysql-connector-python==8.0.27
mysqlclient==2.1.0
networkx==3.0
nibabel==5.0.0
nipype==1.8.5
numpy==1.23.3
oauthlib==3.2.2
opencv-contrib-python==4.7.0.68
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.0
pathlib==1.0.1
Pillow==9.1.0
platformdirs==2.5.2
protobuf==3.19.6
prov==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
pydot==1.4.2
pygame==2.1.2
PyGetWindow==0.0.9
PyJWT==2.5.0
PyMsgBox==1.0.9
PyMuPDF==1.21.1
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2020.1
pyusb==1.2.1
pywhatkit==5.3
pyxnat==1.5
rdflib==6.2.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.0
simplejson==3.18.1
six==1.15.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
starlette==0.23.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.2.0
tqdm==4.64.1
traits==6.3.2
typing-extensions==4.4.0
urllib3==1.26.9
uvicorn==0.20.0
virtualenv==20.14.1
Werkzeug==2.2.2
wikipedia==1.4.0
wrapt==1.14.1
WTForms==3.0.1
zipp==3.11.0
